@@ -1,22 +1,26 @@
 """Global fixtures for dlink_dchs150_hass integration."""
+
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, Any
 from unittest.mock import patch
 
 import pytest
-
-from typing import Dict
-
 from pytest_homeassistant_custom_component.common import MockConfigEntry
 
+if TYPE_CHECKING:
+    from collections.abc import Generator
+
 from custom_components.dchs150_motion.const import (
-    DOMAIN,
-    CONF_HOST,
-    CONF_PIN,
-    CONF_INTERVAL,
     CONF_BACKOFF,
-    CONF_SENSITIVITY,
-    CONF_OP_STATUS,
-    CONF_NICK_NAME,
     CONF_DESCRIPTION,
+    CONF_HOST,
+    CONF_INTERVAL,
+    CONF_NICK_NAME,
+    CONF_OP_STATUS,
+    CONF_PIN,
+    CONF_SENSITIVITY,
+    DOMAIN,
 )
 
 pytest_plugins = "pytest_homeassistant_custom_component"
@@ -26,10 +30,11 @@ pytest_plugins = "pytest_homeassistant_custom_component"
 # notifications. These calls would fail without this fixture since the persistent_notification
 # integration is never loaded during a test.
 @pytest.fixture(name="skip_notifications", autouse=True)
-def skip_notifications_fixture():
+def skip_notifications_fixture() -> Generator[Any]:
     """Skip notification calls."""
-    with patch("homeassistant.components.persistent_notification.async_create"), patch(
-        "homeassistant.components.persistent_notification.async_dismiss"
+    with (
+        patch("homeassistant.components.persistent_notification.async_create"),
+        patch("homeassistant.components.persistent_notification.async_dismiss"),
     ):
         yield
 
@@ -37,10 +42,10 @@ def skip_notifications_fixture():
 # This fixture, when used, will result in calls to async_get_data to return None. To have the call
 # return a value, we would add the `return_value=<VALUE_TO_RETURN>` parameter to the patch call.
 @pytest.fixture(name="bypass_get_data")
-def bypass_get_data_fixture():
+def bypass_get_data_fixture() -> Generator[Any]:
     """Skip calls to get data from API."""
     with patch(
-        "custom_components.dlink_dchs150_hass.DlinkDchHassApiClient.async_get_data"
+        "custom_components.dlink_dchs150_hass.DlinkDchHassApiClient.async_get_data",
     ):
         yield
 
@@ -48,7 +53,7 @@ def bypass_get_data_fixture():
 # In this fixture, we are forcing calls to async_get_data to raise an Exception. This is useful
 # for exception handling.
 @pytest.fixture(name="error_on_get_data")
-def error_get_data_fixture():
+def error_get_data_fixture() -> Generator[Any]:
     """Simulate error when retrieving data from API."""
     with patch(
         "custom_components.dlink_dchs150_hass.DlinkDchHassApiClient.async_get_data",
@@ -58,13 +63,14 @@ def error_get_data_fixture():
 
 
 @pytest.fixture(name="config_data")
-def config_data() -> Dict[str, str]:
+def config_data() -> dict[str, str]:
+    """Create a mock config data."""
     return {CONF_HOST: "10.1.1.1", CONF_PIN: "123456"}
 
 
 @pytest.fixture(name="config_entry")
 def config_entry() -> MockConfigEntry:
-    """Create a mock config entry"""
+    """Create a mock config entry."""
     return MockConfigEntry(
         domain=DOMAIN,
         data=config_data(),
@@ -73,8 +79,8 @@ def config_entry() -> MockConfigEntry:
 
 
 @pytest.fixture(name="config_options")
-def config_options():
-    """Create a mock set of options"""
+def config_options() -> dict[str, int | bool | str]:
+    """Create a mock set of options."""
     return {
         CONF_INTERVAL: 1,
         CONF_BACKOFF: 30,
