@@ -4,10 +4,14 @@ Custom integration to integrate dlink_dchs150_hass with Home Assistant.
 For more details about this integration, please refer to
 https://github.com/updrytwist/dlink-dchs150-hass
 """
+
 import logging
 
+__version__ = "0.2.0-dev0"
+
+__all__: list[str] = []
+
 from homeassistant.config_entries import ConfigEntry
-from homeassistant.core import Config
 from homeassistant.core import HomeAssistant
 
 from .hass_integration import HassIntegration
@@ -15,12 +19,12 @@ from .hass_integration import HassIntegration
 _LOGGER: logging.Logger = logging.getLogger(__package__)
 
 
-async def async_setup(hass: HomeAssistant, config: Config):
-    """Setup using YAML not supported . . ."""
-    return await HassIntegration.async_setup(hass, config)
+async def async_setup(hass: HomeAssistant) -> bool:
+    """Perform setup.  Setup using YAML not supported . . ."""
+    return await HassIntegration.async_setup(hass)
 
 
-async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry):
+async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     """Perform initial setup."""
     return await HassIntegration.async_setup_entry(hass, entry)
 
@@ -32,4 +36,4 @@ async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
 
 async def async_reload_entry(hass: HomeAssistant, entry: ConfigEntry) -> None:
     """Reload config entry."""
-    return await HassIntegration.async_reload_entry(hass, entry)
+    await HassIntegration.async_reload_entry(hass, entry)
