@@ -1,21 +1,20 @@
 """Test dlink_dchs150_hass setup process."""
+
 import pytest
+from homeassistant.core import HomeAssistant
+from homeassistant.exceptions import ConfigEntryNotReady
+
 from custom_components.dchs150_motion import (
     async_reload_entry,
-)
-from custom_components.dchs150_motion import (
     async_setup_entry,
-)
-from custom_components.dchs150_motion import (
     async_unload_entry,
-)
-from custom_components.dchs150_motion.hass_integration import (
-    DlinkDchHassDataUpdateCoordinator,
 )
 from custom_components.dchs150_motion.const import (
     DOMAIN,
 )
-from homeassistant.exceptions import ConfigEntryNotReady
+from custom_components.dchs150_motion.hass_integration import (
+    DlinkDchHassDataUpdateCoordinator,
+)
 
 
 # We can pass fixtures as defined in conftest.py to tell pytest to use the fixture
@@ -23,14 +22,18 @@ from homeassistant.exceptions import ConfigEntryNotReady
 # Home Assistant using the pytest_homeassistant_custom_component plugin.
 # Assertions allow you to verify that the return value of whatever is on the left
 # side of the assertion matches with the right side.
-async def test_setup_unload_and_reload_entry(hass, bypass_get_data, config_entry):
+async def test_setup_unload_and_reload_entry(
+    hass: HomeAssistant,
+    bypass_get_data,  # noqa: ANN001
+    config_entry,  # noqa: ANN001
+) -> None:
     """Test entry setup and unload."""
-
     # Set up the entry and assert that the values set during setup are where we expect
     # them to be. Because we have patched the DlinkDchs150HassDataUpdateCoordinator.async_get_data
     # call, no code from custom_components/dlink_dchs150_hass/api.py actually runs.
     assert await async_setup_entry(hass, config_entry)
-    assert DOMAIN in hass.data and config_entry.entry_id in hass.data[DOMAIN]
+    assert DOMAIN in hass.data
+    assert config_entry.entry_id in hass.data[DOMAIN]
     assert (
         type(hass.data[DOMAIN][config_entry.entry_id])
         is DlinkDchHassDataUpdateCoordinator
@@ -38,7 +41,8 @@ async def test_setup_unload_and_reload_entry(hass, bypass_get_data, config_entry
 
     # Reload the entry and assert that the data from above is still there
     assert await async_reload_entry(hass, config_entry) is None
-    assert DOMAIN in hass.data and config_entry.entry_id in hass.data[DOMAIN]
+    assert DOMAIN in hass.data
+    assert config_entry.entry_id in hass.data[DOMAIN]
     assert (
         type(hass.data[DOMAIN][config_entry.entry_id])
         is DlinkDchHassDataUpdateCoordinator
@@ -49,9 +53,12 @@ async def test_setup_unload_and_reload_entry(hass, bypass_get_data, config_entry
     assert config_entry.entry_id not in hass.data[DOMAIN]
 
 
-async def test_setup_entry_exception(hass, error_on_get_data, config_entry):
+async def test_setup_entry_exception(
+    hass: HomeAssistant,
+    error_on_get_data,  # noqa: ANN001
+    config_entry,  # noqa: ANN001
+) -> None:
     """Test ConfigEntryNotReady when API raises an exception during entry setup."""
-
     # In this case we are testing the condition where async_setup_entry raises
     # ConfigEntryNotReady using the `error_on_get_data` fixture which simulates
     # an error.
